requirements.txt
torch
torchvision
Pillow
torchsummary
numpy
scikit-learn
seaborn
matplotlib
